Module Slurp_cohttp

Slurp Cohttp server definitions.

Quickstart

To use slurp with the cohttp backend, you can simply call the slurp-cohttp binary coming with the so called package. However, if you need to tune the server implementation, you can use the server extensible definition inherited from Clim.cli.

Extension

Most of the time, the server is launched together with other services that may already bind the slurp default option names. To avoid any conflict, the server CLI is defined from Clim.cli and option names are overloadable methods.

The parameters names are defined with string lists following the Cmdliner convention: short names are one char strings and others strings are long names. For example, if you want to rename the "--port" option, simply reinstanciate the server definition with a custom configuration:

open Slurp_cohttp

let my_server = object
  inherit cohttp
  method! port_names = ["my-port"]
end

let _ = Lwt_main.run my_server#run

the other extension oftenly used is adding new options for various services added with the custom REST API. Adding options is allowed by the Clim interface. For example, we can add the option foo simply by registering it with:

let my_server = object
   ...
   val foo = 42
   initializer
     self#arg Clim.(value @@ opt
      ~doc:"My awesome option"
      int
      42
      ["foo"]) (fun v -> foo <- v)
   ...
end

foo will then contain the foo option value at runtime.

More subtle extensions are also available by redefining the My_server.command value to add some custom behavior to the resulting binary. For example:

let my_server = object(self)
  inherit cohttp as super
  ...
  method! entrypoint () =
     Lwt.join [
       super#entrypoint ();
       (fun () -> print_endline "Hello there!"; Lwt.return_unit);
     ]
  ...

will add the trace "Hello there!" on server run.

API

class cohttp : object ... end

A customizable SLURP cohttp definition.